refactor(generator): extract PermissionFeatureChecks
